move preprogrammed steps to examples/locomotion; fix typo
