--- flygym/examples/__init__.py
@@ -1,1 +1,0 @@
-from .common import PreprogrammedSteps, plot_time_series_multi_legs
