+++ flygym/examples/locomotion/__init__.py
@@ -1,3 +1,4 @@
+from .steps import PreprogrammedSteps
 from .cpg_controller import CPGNetwork
 from .rule_based_controller import RuleBasedController
 from .turning_controller import HybridTurningController
